ARUHA-681 Don't fail if topic does not exist
This can happen, for example if a timeline is marked as deleted

+++ src/main/java/org/zalando/nakadi/service/timeline/TimelineService.java
@@ -236,7 +236,9 @@ public class TimelineService {
         final List<Timeline> timelines = getTimelines(eventTypeName);
         for (final Timeline timeline : timelines) {
             final TopicRepository topicRepository = getTopicRepository(timeline);
-            topicRepository.deleteTopic(timeline.getTopic());
+            if (topicRepository.topicExists(timeline.getTopic())) {
+                topicRepository.deleteTopic(timeline.getTopic());
+            }
             timelineDbRepository.deleteTimeline(timeline.getId());
         }
     }
